requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
aiosqlite==0.20.0
aisqlite==1.1.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
docutils==0.21.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jiter==0.8.0
keyring==25.5.0
magic-filter==1.0.12
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
multidict==6.1.0
nh3==0.2.19
openai==1.56.2
packaging==24.2
pkginfo==1.12.0
propcache==0.2.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
setuptools==75.6.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.67.1
twine==6.0.1
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.38.4
yarl==1.18.3
